refactor: Rename handler.go to teleport_event_handler.go
이후에 Slack과 Teleport에 대한 처리를 진행할 때 handler.go에서 모든 핸들링을 진행하면 코드의 가독성이 떨어져 분리하였습니다.
